requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-socketio
flask-limiter
flask-babelplus
flask-mailman
flask-security-too
flask-wtf
flask-caching
flask-smorest
flask-jwt-extended

pymysql
python-dotenv
authlib
gunicorn
eventlet
requests
redis

colorama